Fixed validate/get schema button for streaming source
Moved Spark related classes out of the source and into a separate
class. This allows the CDAP validation service to instantiate
the source for validation/schema purposes without hitting the
Spark class not found exceptions.
